New camera  window, checklist improvements, connection status improvements
The camerea feed can now be opened in a new window and resized from there if the drivers wish for this. This still needs testing.

Some improvements have been made to saving, namely if the save directory doesn't exist the program uses the installed directory.

Multiple checklist items can be deleted at once.

The connection screen now features a disconnect button if the robot is connected. Needs testing.