Fixed gitignore some more and improved warning system and checklist
Checklist warns the warning system of when its items are completely finished better now, including when items are deleted.

Also made changes to the way warnings are handled from the robot. Warnings from the robot have their own list to which they are added as well as the regular list to prevent the robot from modifying warnings that it didn't set.