Fixed error in robot logging and improved connection notifier
The robot logging system didn't check for valid log names before it used them, which may be the reason why they didn't save during
competition.

The connection status is now faster to update because there is a callback function that handles it for the connectionUI instead of relying
on the regular event and having it be a little slow.